requirements.txt
aiohttp==3.6.2
amqp==2.5.2
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828452693/work
asgiref==3.2.7
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.1.4
blinker==1.4
bokeh==2.0.2
celery==4.4.2
celery-singleton==0.2.0
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
cmarkgfm==0.4.2
colorama==0.4.3
colorlog==4.2.1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.3
django-admin-honeypot==1.1.0
django-allauth==0.41.0
django-celery-results==1.2.1
django-crispy-forms==1.9.1
django-otp==0.9.1
django-redis==4.11.0
django-redis-cache==2.1.1
django-tables2==2.3.1
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
fastlogging==0.9.4
future==0.18.2
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
hiredis==1.0.1
idna==2.9
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206598566/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi==0.17.0
Jinja2==2.11.2
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148959479/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
kombu==4.6.8
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846017414/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
multidict==4.7.3
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914821128/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
notebook @ file:///tmp/build/80754af9/notebook_1601501580008/work
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.0.3
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pika==1.1.0
Pillow==7.1.2
pkginfo==1.5.0.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
psycopg2cffi==2.8.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1605560620615/work/dist/ptyprocess-0.6.0-py2.py3-none-any.whl
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
pylibmc==1.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141707582/work
PySocks==1.7.1
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
qrcode==6.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
readme-renderer==24.0
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scipy==1.3.3
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six==1.14.0
SQLAlchemy @ file:///tmp/build/80754af9/sqlalchemy_1608243750780/work
sqlparse==0.3.1
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
tqdm==4.46.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
twine==2.0.0
typing-extensions==3.7.4.1
urllib3==1.25.8
vine==1.3.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
yarl==1.4.2
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
